+++ docs/earthpy.rst
@@ -28,6 +28,14 @@ earthpy.mask module
     :undoc-members:
     :show-inheritance:
 
+earthpy.plot module
+-------------------
+
+.. automodule:: earthpy.plot
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
 earthpy.spatial module
 ----------------------
