+++ earthpy/tests/test_plot.py
@@ -2,8 +2,8 @@
 
 import numpy as np
 import pytest
-import earthpy.spatial as es
 import matplotlib.pyplot as plt
+import earthpy.plot as ep
 
 plt.show = lambda: None
 
@@ -18,43 +18,46 @@ def one_band_3dims():
     return np.random.randint(10, size=(1, 4, 5))
 
 
-""" General functions for matplotlib elements """
+""" plot_bands tests """
 
 
 def test_arr_parameter():
     """Raise an AttributeError if an array is not provided."""
     with pytest.raises(AttributeError):
-        es.plot_bands(arr=(1, 2))
+        ep.plot_bands(arr=(1, 2))
 
 
 def test_num_titles(image_array_2bands):
-    """If a user provides two titles for a single band array, the function
-    should raise an error OR if the title list is a different length than
-    the array it should also raise an errors"""
+    """Test the number of titles.
 
+    If a user provides two titles for a single band array, the function
+    should raise an error OR if the title list is a different length than
+    the array it should also raise an errors.
+    """
     single_band = image_array_2bands[0]
 
     with pytest.raises(ValueError):
-        es.plot_bands(arr=single_band, title=["Title1", "Title2"])
+        ep.plot_bands(arr=single_band, title=["Title1", "Title2"])
     with pytest.raises(ValueError):
-        es.plot_bands(
+        ep.plot_bands(
             arr=image_array_2bands, title=["Title1", "Title2", "Title3"]
         )
 
 
 def test_num_axes(image_array_2bands):
-    """If provided with a 2 band array, plot_bands should return 3 axes by
-    default"""
+    """Test the number of axes.
 
-    f, ax = es.plot_bands(image_array_2bands)
+    If provided with a 2 band array, plot_bands should return 3 axes.
+    """
+    f, ax = ep.plot_bands(image_array_2bands)
     assert len(f.axes) == 3
     plt.close(f)
 
 
 def test_two_plot_title(image_array_2bands):
-    """Test that the default title is provided for a 2 band array plot"""
+    """Test that the default title is provided for a 2 band array plot."""
 
-    f, ax = es.plot_bands(image_array_2bands)
+    f, ax = ep.plot_bands(image_array_2bands)
     ax = f.axes
     num_plts = image_array_2bands.shape[0]
     all_titles = [ax[i].get_title() for i in range(num_plts)]
@@ -63,9 +66,8 @@ def test_two_plot_title(image_array_2bands):
 
 
 def test_custom_plot_title(image_array_2bands):
-    """Test that the custom title is applied for a 2 band array plot"""
-
-    f, ax = es.plot_bands(image_array_2bands, title=["Red Band", "Green Band"])
+    """Test that the custom title is applied for a 2 band array plot."""
+    f, ax = ep.plot_bands(image_array_2bands, title=["Red Band", "Green Band"])
     ax = f.axes
     num_plts = image_array_2bands.shape[0]
     all_titles = [ax[i].get_title() for i in range(num_plts)]
@@ -74,10 +76,12 @@ def test_custom_plot_title(image_array_2bands):
 
 
 def test_single_band_3dims(one_band_3dims):
-    """If you provide a single band array with 3 dimensions (shape[0]==1
-    test that it still plots and only returns a single axis"""
+    """Test single band plot with three dimensions.
 
-    f, ax = es.plot_bands(one_band_3dims)
+    If you provide a single band array with 3 dimensions (shape[0]==1
+    test that it still plots and only returns a single axis.
+    """
+    f, ax = ep.plot_bands(one_band_3dims)
     arr = f.axes[0].get_images()[0].get_array()
     assert arr.ndim == 2
     assert len(f.axes[0].get_images()) == 1
@@ -85,11 +89,13 @@ def test_single_band_3dims(one_band_3dims):
 
 
 def test_single_band_2dims(one_band_3dims):
-    """If you provide a single band array with 3 dimensions (shape[0]==1
-    test that it still plots and only returns a single axis"""
+    """Test single band plot with two dimensions
 
+    If you provide a single band array with 3 dimensions (shape[0] == 1)
+    test that it still plots and only returns a single axis.
+    """
     single_band_2dims = one_band_3dims[0]
-    f, ax = es.plot_bands(single_band_2dims)
+    f, ax = ep.plot_bands(single_band_2dims)
     # Get array from mpl figure
     arr = f.axes[0].get_images()[0].get_array()
     assert arr.ndim == 2
@@ -101,80 +107,77 @@ def test_single_band_2dims(one_band_3dims):
 
 
 def test_colorbar_height(basic_image):
-    """Test that the colorbar ax height matches the image axis height"""
-
+    """Test that the colorbar ax height matches the image axis height."""
     f, ax = plt.subplots(figsize=(5, 5))
     im = ax.imshow(basic_image, cmap="RdYlGn")
-    cb = es.colorbar(im)
-
+    cb = ep.colorbar(im)
     assert cb.ax.get_position().height == im.axes.get_position().height
     plt.close(f)
 
 
 def test_colorbar_raises_value_error():
-    """Test that a non matbplotlib axis object raises an value error"""
-
+    """Test that a non matbplotlib axis object raises an value error."""
     with pytest.raises(AttributeError, match="requires a matplotlib"):
-        es.colorbar(list())
+        ep.colorbar(list())
 
 
 """ Hist tests """
 
 
 def test_num_plot_titles_mismatch_hist(image_array_2bands):
-    """ Raise an error if the number of titles != number of bands """
+    """Raise an error if the number of titles != number of bands."""
     with pytest.raises(ValueError, match="number of plot titles"):
-        es.hist(image_array_2bands, title=["One", "too", "many"])
+        ep.hist(image_array_2bands, title=["One", "too", "many"])
 
 
 def test_num_axes_hist(image_array_2bands, basic_image):
-    """ We expect one AxesSubplot object per pand """
-    f_1, ax_1 = es.hist(basic_image)
+    """We expect one AxesSubplot object per band."""
+    f_1, ax_1 = ep.hist(basic_image)
     assert len(f_1.axes) == 1
-    f_2, ax_2 = es.hist(image_array_2bands)
+    f_2, ax_2 = ep.hist(image_array_2bands)
     assert len(f_2.axes) == 2
 
 
 def test_single_hist_title(basic_image):
-    """ Test that custom titles work for one band hists """
+    """Test that custom titles work for one band hists."""
     custom_title = "Great hist"
-    f, ax = es.hist(basic_image, title=[custom_title])
+    f, ax = ep.hist(basic_image, title=[custom_title])
     assert ax.get_title() == custom_title
 
 
 def test_multiband_hist_title(image_array_2bands):
-    """ Test that custom titles work for multiband hists """
+    """Test that custom titles work for multiband hists."""
     custom_titles = ["Title 1", "Title 2"]
-    f, ax = es.hist(image_array_2bands, title=custom_titles)
+    f, ax = ep.hist(image_array_2bands, title=custom_titles)
     num_plts = image_array_2bands.shape[0]
     assert [f.axes[i].get_title() for i in range(num_plts)] == custom_titles
 
 
 def test_number_of_hist_bins(basic_image):
-    """ Test that the number of bins is customizable """
+    """Test that the number of bins is customizable."""
     n_bins = [1, 10, 100]
     for n in n_bins:
-        f, ax = es.hist(basic_image, bins=n)
+        f, ax = ep.hist(basic_image, bins=n)
         assert n == len(ax.patches)
 
 
 def test_hist_bbox(basic_image):
-    """ Verify that the bbox dimensions are customizable """
-    f, ax = es.hist(basic_image, figsize=(50, 3))
+    """Test that the bbox dimensions are customizable."""
+    f, ax = ep.hist(basic_image, figsize=(50, 3))
     bbox = str(f.__dict__.get("bbox_inches"))
     assert bbox == "Bbox(x0=0.0, y0=0.0, x1=50.0, y1=3.0)"
 
 
 def test_hist_color_single_band(basic_image):
-    """ Check that the color argument works for single bands """
-    f, ax = es.hist(basic_image, colors=["red"])
+    """Check that the color argument works for single bands."""
+    f, ax = ep.hist(basic_image, colors=["red"])
     facecolor = ax.patches[0].__dict__.get("_original_facecolor")
     assert np.array_equal(facecolor, np.array([1.0, 0.0, 0.0, 1.0]))
 
 
 def test_hist_color_multi_band(image_array_2bands):
-    """ Verify that multiple colors work for multiband images """
-    f, ax = es.hist(image_array_2bands, colors=["red", "blue"])
+    """Test that multiple colors work for multiband images."""
+    f, ax = ep.hist(image_array_2bands, colors=["red", "blue"])
     colors = [a.patches[0].__dict__.get("_original_facecolor") for a in ax]
     expected_colors = [
         np.array([1.0, 0.0, 0.0, 1.0]),
@@ -185,8 +188,8 @@ def test_hist_color_multi_band(image_array_2bands):
 
 
 def test_hist_number_of_columns(image_array_2bands):
-    """ Verify that the col argument changes the number of columns """
+    """Test that the col argument changes the number of columns."""
     number_of_columns = [1, 2]
     for n in number_of_columns:
-        f, ax = es.hist(image_array_2bands, cols=n)
+        f, ax = ep.hist(image_array_2bands, cols=n)
         assert [a.numCols for a in ax] == [n] * 2
